+++ game/src/main/kotlin/gg/rsmod/game/message/handler/OpLoc3Handler.kt
@@ -1,15 +1,18 @@
 package gg.rsmod.game.message.handler
 
+import gg.rsmod.game.action.GroundItemPathAction
 import gg.rsmod.game.action.ObjectPathAction
 import gg.rsmod.game.message.MessageHandler
 import gg.rsmod.game.message.impl.OpLoc3Message
 import gg.rsmod.game.model.EntityType
 import gg.rsmod.game.model.Tile
 import gg.rsmod.game.model.World
+import gg.rsmod.game.model.attr.INTERACTING_GROUNDITEM_ATTR
 import gg.rsmod.game.model.attr.INTERACTING_OBJ_ATTR
 import gg.rsmod.game.model.attr.INTERACTING_OPT_ATTR
 import gg.rsmod.game.model.entity.Client
 import gg.rsmod.game.model.entity.GameObject
+import gg.rsmod.game.model.entity.GroundItem
 import gg.rsmod.game.model.entity.Player
 import gg.rsmod.game.model.priv.Privilege
 import java.lang.ref.WeakReference
@@ -20,7 +23,7 @@ import java.lang.ref.WeakReference
 class OpLoc3Handler : MessageHandler<OpLoc3Message> {
 
     override fun handle(client: Client, world: World, message: OpLoc3Message) {
-        /*
+        /**
          * If tile is too far away, don't process it.
          */
         val tile = Tile(message.x, message.z, client.tile.height)
@@ -28,33 +31,28 @@ class OpLoc3Handler : MessageHandler<OpLoc3Message> {
             return
         }
 
-        /*
-         * If player can't move, we don't do anything.
-         */
-        if (!client.lock.canMove()) {
+        if (!client.lock.canGroundItemInteract()) {
             return
         }
 
-        /*
+        log(client, "Ground Item action 3: item=%d, x=%d, z=%d, movement=%d", message.item, message.x, message.z, message.movementType)
+
+        /**
          * Get the region chunk that the object would belong to.
          */
         val chunk = world.chunks.getOrCreate(tile)
-        val obj = chunk.getEntities<GameObject>(tile, EntityType.STATIC_OBJECT, EntityType.DYNAMIC_OBJECT).firstOrNull { it.id == message.id } ?: return
+        val item = chunk.getEntities<GroundItem>(tile, EntityType.GROUND_ITEM).firstOrNull { it.item == message.item && it.canBeViewedBy(client) } ?: return
 
-        log(client, "Object action 3: id=%d, x=%d, z=%d, movement=%d", message.id, message.x, message.z, message.movementType)
+        if (message.movementType == 1 && world.privileges.isEligible(client.privilege, Privilege.ADMIN_POWER)) {
+            client.moveTo(item.tile)
+        }
 
-        client.stopMovement()
         client.closeInterfaceModal()
         client.interruptQueues()
         client.resetInteractions()
 
-        if (message.movementType == 1 && world.privileges.isEligible(client.privilege, Privilege.ADMIN_POWER)) {
-            val def = obj.getDef(world.definitions)
-            client.moveTo(world.findRandomTileAround(obj.tile, radius = 1, centreWidth = def.width, centreLength = def.length) ?: obj.tile)
-        }
-
         client.attr[INTERACTING_OPT_ATTR] = 3
-        client.attr[INTERACTING_OBJ_ATTR] = WeakReference(obj)
-        client.executePlugin(ObjectPathAction.objectInteractPlugin)
+        client.attr[INTERACTING_GROUNDITEM_ATTR] = WeakReference(item)
+        client.executePlugin(GroundItemPathAction.walkPlugin)
     }
 }
