+++ game/src/main/kotlin/gg/rsmod/game/message/handler/OpObj3Handler.kt
@@ -1,14 +1,17 @@
 package gg.rsmod.game.message.handler
 
 import gg.rsmod.game.action.GroundItemPathAction
+import gg.rsmod.game.action.ObjectPathAction
 import gg.rsmod.game.message.MessageHandler
 import gg.rsmod.game.message.impl.OpObj3Message
 import gg.rsmod.game.model.EntityType
 import gg.rsmod.game.model.Tile
 import gg.rsmod.game.model.World
 import gg.rsmod.game.model.attr.INTERACTING_GROUNDITEM_ATTR
+import gg.rsmod.game.model.attr.INTERACTING_OBJ_ATTR
 import gg.rsmod.game.model.attr.INTERACTING_OPT_ATTR
 import gg.rsmod.game.model.entity.Client
+import gg.rsmod.game.model.entity.GameObject
 import gg.rsmod.game.model.entity.GroundItem
 import gg.rsmod.game.model.entity.Player
 import gg.rsmod.game.model.priv.Privilege
@@ -20,7 +23,7 @@ import java.lang.ref.WeakReference
 class OpObj3Handler : MessageHandler<OpObj3Message> {
 
     override fun handle(client: Client, world: World, message: OpObj3Message) {
-        /**
+        /*
          * If tile is too far away, don't process it.
          */
         val tile = Tile(message.x, message.z, client.tile.height)
@@ -28,28 +31,33 @@ class OpObj3Handler : MessageHandler<OpObj3Message> {
             return
         }
 
-        if (!client.lock.canGroundItemInteract()) {
+        /*
+         * If player can't move, we don't do anything.
+         */
+        if (!client.lock.canMove()) {
             return
         }
 
-        log(client, "Ground Item action 3: item=%d, x=%d, z=%d, movement=%d", message.item, message.x, message.z, message.movementType)
-
-        /**
+        /*
          * Get the region chunk that the object would belong to.
          */
         val chunk = world.chunks.getOrCreate(tile)
-        val item = chunk.getEntities<GroundItem>(tile, EntityType.GROUND_ITEM).firstOrNull { it.item == message.item && it.canBeViewedBy(client) } ?: return
+        val obj = chunk.getEntities<GameObject>(tile, EntityType.STATIC_OBJECT, EntityType.DYNAMIC_OBJECT).firstOrNull { it.id == message.id } ?: return
 
-        if (message.movementType == 1 && world.privileges.isEligible(client.privilege, Privilege.ADMIN_POWER)) {
-            client.moveTo(item.tile)
-        }
+        log(client, "Object action 3: id=%d, x=%d, z=%d, movement=%d", message.id, message.x, message.z, message.movementType)
 
+        client.stopMovement()
         client.closeInterfaceModal()
         client.interruptQueues()
         client.resetInteractions()
 
+        if (message.movementType == 1 && world.privileges.isEligible(client.privilege, Privilege.ADMIN_POWER)) {
+            val def = obj.getDef(world.definitions)
+            client.moveTo(world.findRandomTileAround(obj.tile, radius = 1, centreWidth = def.width, centreLength = def.length) ?: obj.tile)
+        }
+
         client.attr[INTERACTING_OPT_ATTR] = 3
-        client.attr[INTERACTING_GROUNDITEM_ATTR] = WeakReference(item)
-        client.executePlugin(GroundItemPathAction.walkPlugin)
+        client.attr[INTERACTING_OBJ_ATTR] = WeakReference(obj)
+        client.executePlugin(ObjectPathAction.objectInteractPlugin)
     }
 }
