+++ game/src/main/kotlin/gg/rsmod/game/message/impl/OpLoc1Message.kt
@@ -5,4 +5,4 @@ import gg.rsmod.game.message.Message
 /**
  * @author Tom <rspsmods@gmail.com>
  */
-data class OpLoc1Message(val id: Int, val x: Int, val z: Int, val movementType: Int) : Message
+data class OpLoc1Message(val item: Int, val x: Int, val z: Int, val movementType: Int) : Message
