+++ game/src/main/kotlin/gg/rsmod/game/message/impl/OpLoc4Message.kt
@@ -5,4 +5,4 @@ import gg.rsmod.game.message.Message
 /**
  * @author Tom <rspsmods@gmail.com>
  */
-data class OpLoc4Message(val id: Int, val x: Int, val z: Int, val movementType: Int) : Message
+data class OpLoc4Message(val item: Int, val x: Int, val z: Int, val movementType: Int) : Message
