+++ game/src/main/kotlin/gg/rsmod/game/message/impl/OpObj1Message.kt
@@ -5,4 +5,4 @@ import gg.rsmod.game.message.Message
 /**
  * @author Tom <rspsmods@gmail.com>
  */
-data class OpObj1Message(val item: Int, val x: Int, val z: Int, val movementType: Int) : Message
+data class OpObj1Message(val id: Int, val x: Int, val z: Int, val movementType: Int) : Message
