+++ game/src/main/kotlin/gg/rsmod/game/message/impl/OpObj2Message.kt
@@ -5,4 +5,4 @@ import gg.rsmod.game.message.Message
 /**
  * @author Tom <rspsmods@gmail.com>
  */
-data class OpObj2Message(val item: Int, val x: Int, val z: Int, val movementType: Int) : Message
+data class OpObj2Message(val id: Int, val x: Int, val z: Int, val movementType: Int) : Message
